REFACT: 오브젝트와 빛, 유효성 검사 파일 분리
