+++ src/parse/is_valid/is_valid_elements_object.c
@@ -1,22 +1,5 @@
 #include "../../../include/parse.h"
 
-t_is_return	is_valid_a_element(char **elements)
-{
-	char	*last_element;
-
-	last_element = elements[count_elements(elements) - 1];
-	last_element[ft_strlen(last_element) - 1] = '\0';
-	if (count_elements(elements) != 3)
-		return (NO);
-	if (is_valid_double_format(elements[1]) == NO)
-		return (NO);
-	if (is_between(0.0, 1.0, ft_atod(elements[1])) == NO)
-		return (NO);
-	if (is_rgb(elements[2]) == NO)
-		return (NO);
-	return (YES);
-}
-
 t_is_return	is_valid_c_element(char **elements)
 {
 	char	*last_element;
@@ -32,25 +15,6 @@ t_is_return	is_valid_c_element(char **elements)
 	if (is_valid_double_format(last_element) == NO)
 		return (NO);
 	if (is_between(0.0, 180.0, ft_atod(last_element)) == NO)
-		return (NO);
-	return (YES);
-}
-
-t_is_return	is_valid_l_element(char **elements)
-{
-	char	*last_element;
-
-	last_element = elements[count_elements(elements) - 1];
-	last_element[ft_strlen(last_element) - 1] = '\0';
-	if (count_elements(elements) != 4)
-		return (NO);
-	if (is_xyz_coordinate(elements[1]) == NO)
-		return (NO);
-	if (is_valid_double_format(elements[2]) == NO)
-		return (NO);
-	if (is_between(0.0, 1.0, ft_atod(elements[2])) == NO)
-		return (NO);
-	if (is_rgb(elements[3]) == NO)
 		return (NO);
 	return (YES);
 }
